requirements.txt
matplotlib
wordcloud
requests
Pillow
numpy
python-dotenv
